Move getSelection to generic browserEnvironment/helpers.
