+++ src/e2e/browserEnvironment/helpers/getSelection.ts
@@ -1,5 +1,5 @@
 /* eslint-disable fp/no-get-set */
-import { BrowserEnvironment } from '../../test-types'
+import { BrowserEnvironment } from '../types'
 
 /** Returns a proxy getSelection object with async getters for selection properties. WebdriverIO cannot retrieve the entire selection object so we delegate to asynchronous getters. Top-level selection properties (e.g. .focusNode) are compatible with nested properties (e.g. .focusNode.textContent) are compatible with parent properties by defining ad hoc getters on promises.
  *
