+++ src/e2e/browserEnvironment/types.ts
