+++ src/e2e/iOS/helpers/index.ts
@@ -14,7 +14,7 @@ import initSession from './initSession'
 import isKeyboardShown from './isKeyboardShown'
 import hideKeyboardByTappingDone from './hideKeyboardByTappingDone'
 import paste from './paste'
-import getSelection from './getSelection'
+import getSelection from '../../browserEnvironment/helpers/getSelection'
 import tapReturnKey from './tapReturnKey'
 import tap from './tap'
 import type from './type'
