+++ src/e2e/puppeteer/helpers/getSelection.ts
@@ -1,6 +1,6 @@
 import { Page } from 'puppeteer'
 import asBrowserEnvironment from './asBrowserEnvironment'
-import getSelectionWebdriver from '../../iOS/helpers/getSelection'
+import getSelectionWebdriver from '../../browserEnvironment/helpers/getSelection'
 
 /** Returns a proxy getSelection object with async getters for selection properties. */
 const getSelection = (page: Page) => getSelectionWebdriver(asBrowserEnvironment(page))
